Move custom serializers to EventStream module
