+++ lib/event_stream.rb
@@ -1,0 +1,8 @@
+require 'event_stream/project_serializer'
+require 'event_stream/workflow_serializer'
+require 'event_stream/user_serializer'
+require 'event_stream/subject_serializer'
+require 'event_stream/classification_serializer'
+
+module EventStream
+end
