+++ lib/event_stream/project_serializer.rb
@@ -1,0 +1,5 @@
+module EventStream
+  class ProjectSerializer < ActiveModel::Serializer
+    attributes :id, :name, :created_at
+  end
+end
