+++ lib/event_stream/subject_serializer.rb
@@ -1,0 +1,5 @@
+module EventStream
+  class SubjectSerializer < ActiveModel::Serializer
+    attributes :id, :metadata, :created_at, :updated_at
+  end
+end
