+++ lib/event_stream/user_serializer.rb
@@ -1,0 +1,5 @@
+module EventStream
+  class UserSerializer < ActiveModel::Serializer
+    attributes :id, :login
+  end
+end
