+++ lib/event_stream/workflow_serializer.rb
@@ -1,0 +1,5 @@
+module EventStream
+  class WorkflowSerializer < ActiveModel::Serializer
+    attributes :id, :created_at
+  end
+end
