--- lib/kafka_project_serializer.rb
@@ -1,3 +1,0 @@
-class KafkaProjectSerializer < ActiveModel::Serializer
-  attributes :id, :name, :created_at
-end
