--- lib/kafka_subject_serializer.rb
@@ -1,3 +1,0 @@
-class KafkaSubjectSerializer < ActiveModel::Serializer
-  attributes :id, :metadata, :created_at, :updated_at
-end
