--- lib/kafka_user_serializer.rb
@@ -1,3 +1,0 @@
-class KafkaUserSerializer < ActiveModel::Serializer
-  attributes :id, :login
-end
