--- lib/kafka_workflow_serializer.rb
@@ -1,3 +1,0 @@
-class KafkaWorkflowSerializer < ActiveModel::Serializer
-  attributes :id, :created_at
-end
